Increase exception specificity in getPackage
The package resolver calls getValueOrThrow in its getPackage method.
This call can throw an exception, and a year ago, the exception could
be multiple different subtypes of NoSuchThingException. Subsequently,
we made it so that only NoSuchPackageException could be thrown, but
the exception type specified in this call site remained the more
general supertype.

--
MOS_MIGRATED_REVID=88154358