Added Eclipse project configuration for Bazel
This add the eclipse project configuration. To use it, first do a
`./bootstrap_test.sh all` then import the Eclipse project in the Bazel
directory.
On my mac using Eclipse Luna configured with a JDK 8 it still displays
errors on truth's isEmpty() method on collections. We Might need to
tweak a bit the java configuration for it to work nicely.

--
MOS_MIGRATED_REVID=88321458