Simplify BuildConfiguration to make it cheaper to create more instances.
We want to make it so configured targets can only see the configuration
fragments they have declared, which requires creating a customized
configuration for every configured target.

Therefore, this change moves the directories to a separate object, so
configurations that have the same output directories can share it.

Also change getBinFragment and getGenfilesFragment to call the corresponding
Root.getExecPath() on the fly. However, these are called _very_ often, so at
the same time change Root to cache the exec path.

Instead of keeping a map of all executables, only keep the shell executable,
which is the only value that was actually used anywhere.

Remove the cache key, but keep the short cache key. The cache key was only
used to check for duplicate cache keys in the BuildConfigurationCollection,
which is no longer necessary as we don't use the cache key for anything else
anymore. We should instead check for output directory conflicts, but we can't
do that quite yet (I'm working on it).

Also update the documentation in a couple of places.

--
MOS_MIGRATED_REVID=88712178