Remove BuilderUtils#rethrowCause(...)'s vestigial bits
Rethrowing cause made sense a while ago, but now the only usages of this
method have a direct handle on the exception they want dealt with and
wrap the input with "new Exception(...)". Instead just pass the exception
along directly.

Also modified the comment in the unexpected case- the class referenced
does not exist anymore, from what I can tell, and thus the comment was
misleading.

--
MOS_MIGRATED_REVID=90623740