Simplify AnalysisEnvironment creation.
Instead of computing flags based on the configuration in two places, we pass
in the configuration and let the common code figure it out.

Also change it so each CT only depends on the build info collection for its
specific configuration. This should be safe, as the API doesn't allow passing
a configuration in - it instead takes ruleContext.getConfiguration, which
should be consistent with the dependency declarations in SkyframeBuildView in
all cases.

--
MOS_MIGRATED_REVID=91819844