Make SkyKey cached hash code transient
SkyKey argument hashcodes are not stable, because they're composed
of values whose hashcodes may not be stable, such as Java enums.
Therefore a SkyKey's own hashcode isn't stable. And this is fine,
but if we try to serialize and then deserialize the SkyKey with its
cached hashcode, the deserialized SkyKey's cached hashcode won't
match a normally constructed SkyKey, despite them being equal.

Because a SkyKey will deserialize with a default value of 0 for its
cached hashcode, this change also introduces a transient boolean
guard to note whether the correct hashcode has been calculated and
cached.

--
MOS_MIGRATED_REVID=91985674