Ensure extension's bundle minimum OS version is at least 8.0.
This is required because Apple does not accept extensions whose plist
has a minimum OS value lower than 8.0, or whose code is compiled at less
than 8.0, even if it is included in an application that is compiled for,
and marked compatible with, a lower OS version.

Note that I didn't make this an attribute for now as there are no use
cases for setting the value to anything but 8.0. If we ever need to
make this user-configurable, a value can easily be set.

Second submission attempt with small additional logic to distinguish xcode
projects across configurations.

--
MOS_MIGRATED_REVID=92267493