Fix JNI on 32-bit systems
The stat struct's size field is off_t, which varies in size between 32- and
64-bit systems. The size field in FileStatus.java is a long (which in Java
is always 64 bits) so the size is getting initialized incorrectly.

I can't think of a good way to add this to our integration tests, but I've
tested in manually on a VM.

Fixes #174.

--
MOS_MIGRATED_REVID=93038537