Add a --javac_extdir flag to enable configuring the javac -extdirs flag
The -extdirs flag sets a search path to the directory containing the JDK's
extension classes, and should always be set explicitly when cross-compiling. If
it is unset, the ext directory of the host JDK will be used instead.

javac requires that -extdirs be a list of directories. The blaze --javac_extdir
flag takes the label of a filegroup with the 'path' attribute set, and passes
the directory path through to JavaBuilder.

RELNOTES: N/A

--
MOS_MIGRATED_REVID=93147656