Get the code cache directory for stub application from the Context instead of hard-coding it and move the incremental deployment directory from /sdcard/incrementaldeployment to /data/local/tmp/incrementaldeployment.
The first is necessary because if the app is not run under the default user profile (e.g. Android For Work), the default location is not writable, thus, sadness, the second is necessary because /sdcard is not accessible from the non-default profile.

This also makes it possible to eventually install .so files there, because the +x bit can be set under /data/local/tmp (unlike under /sdcard).

--
MOS_MIGRATED_REVID=93287264