Make local TestStrategy bits rely less on runfiles manifests
TestStrategy#getLocalRunfilesDirectory(...) was using manifest names
to infer if the runfiles symlinks were created by comparing the name
of manifests. This is a bit confusing and adds unnecessary reliance
on manifests which we'd like to reduce. Instead get whether or not
we created the symlinks from the boolean that determines it.

--
MOS_MIGRATED_REVID=93414150