Make ObjcBundleLibrary use the "families" attribute.
Currently the ObjcBundleLibrary#create() function creates a
BundleSupport object with the default "iphone" target family.
An ObjcBundleLibrary is a descendant of a Bundle, which has
a "families" attribute, so the information is available.
Currently, ObjcBundleLibrary will unconditionally use "iphone"
as the target device family.

This commit updates ObjcBundleLibrary to use the "families"
attribute. It does this by making BundleSupport
control the "families" attribute, and then provide access
to it for the other classes that use it, like
ObjcBundleLibrary. The other classes are changed to
get their target device family information from their
BundleSupport objects.

--
Change-Id: I4a5be2849315dabfe7f2975c006a2092a56027af
Reviewed-on: https://bazel-review.googlesource.com/#/c/1260/3
MOS_MIGRATED_REVID=93456579