Deduplicate include directories when include scanning.
Duplicate include directories were always possible, but are now much more likely because CppCompileAction#getSystemIncludeDirs will overlap with CppCompileAction#getBuiltInIncludeDirs. Note that it is possible that neither is a subset of the other -- raw specification of C++ options in the CROSSTOOL or elsewhere will be in getSystemIncludeDirs and not in getBuiltInIncludeDirs. Similarly, the include directory under the sysroot will be in getBuiltInIncludeDirs but not in getSystemIncludeDirs.

Duplicate directories are bad because if an #include_next is specified in a file inside a directory listed twice, the scanner may just find that file again in the next occurrence of the directory, as opposed to a different directory listed later.

--
MOS_MIGRATED_REVID=95165367