Expand JavaSourceInfoProvider to provide more fine-grained sources.
This will be used by Jack support, which needs to know the different types
of sources as they're handled differently on the Bazel side. Additionally,
it wants to know the difference between source jars the user provided directly
in the sources of the rule, and source jars used to provide sources for the jar
files in the same rule.

--
MOS_MIGRATED_REVID=95324035