Make android_sdk_repository write a proper BUILD file instead of a stub.
Tested by building a trivial Android app manually. The only remaining issue (apart from all this being terribly untested) is that the runfiles path of main_dex_list_creator is not yet determined correctly, but that is only a problem when using legacy multidexing, so it will be good enough for now.

--
MOS_MIGRATED_REVID=95609792