Fixed the deterministic test flakiness
If you already ran the deterministic test with an earlier version of
the depot, the second run would build only once Bazel and use the
previous run log to do the diff, leading to an obvious failure as
the code did change. It was due to an incorrect parameter default
value that pointed to the diff of the previous run, skipping the
first bazel compilation.

--
MOS_MIGRATED_REVID=97399620