Handle errors from DirectoryListingFunction
SkyFunctions that call DirectoryListingFunction directly, as in the
case of RecursiveDirectoryTraversalFunction, or transitively, as in
the case of IncludeParser's call to GlobFunction, had been failing
to handle the exceptions that DirectoryListingFunction can throw.

DirectoryListingFunction can throw InconsistentFilesystemException
directly, but it can also throw IOException and
FileOutsidePackageRootsException because of its call to
DirectoryListingStateFunction without any of its own error handling.

RecursiveFilesystemTraversalFunction also calls
DirectoryListingFunction, but is not yet in use. A follow-up CL will
take care of its error handling needs.

--
MOS_MIGRATED_REVID=97828177