Implement the android_local_tools_repository rule.
This (and its eventual sibling that downloads these from a .zip file) will be used to relax the requirement that Android tools need to be present in every repository where Android builds are done.

Due to lack of any reasonable way to put this in a test, I tested it manually by removing the src and tools/android directories in a workspace and running an Android build.

--
MOS_MIGRATED_REVID=97849649