Move Jack's tools into AndroidSdkProvider.
This also eliminates the JackRule, which is no longer necessary, as its sole
purpose was to allow Jack's tools to be found. This is now part of
AndroidSdkProvider, like all other Android tools.

For now, the new attributes on android_sdk are optional and default to their
old values. In the future, the new attributes will become mandatory with no
defaults, like the other attributes on that rule.

Also fixes a bug where errors found during AndroidCommon.initJava would not
result in a null return from init, previously obscured by the early return
from initJack.

--
MOS_MIGRATED_REVID=98305022