Change the default for CcLibraryHelper.emitLinkActionsIfEmpty to false.
This is renamed from the previous name emitCompileActionsIfEmpty, which was
a misnomer, because it didn't affect compile actions at all.

Update the callers to no longer call the method if not necessary, which leaves
only CcLibrary. CcBinary doesn't hit the link action code path, as it sets the
link type to either DYNAMIC_LIBRARY or EXECUTABLE.

This is in preparation for removing the implicit outputs from cc_library, which
should allow building header-only libraries on MacOS.

--
MOS_MIGRATED_REVID=98927221