Bazel release process scripts
Those scripts are made to handle the soon to be Bazel
release process. It is still missing the CI part that
deploys built artifacts and sends announcement mails.

It uses the RELNOTES: tag (RELNOTES for a simple change,
RELNOTES[NEW] for a new feature, RELNOTES[INC] for an
incompatible change) to create the CHANGELOG.md file.

The main entry point is scripts/release/release.sh
and should be the only script called directly.

--
Change-Id: I6e066fda4d862ce793cc03dc8bdf0b8cb34f7c2c
Reviewed-on: https://bazel-review.googlesource.com/#/c/1580/
MOS_MIGRATED_REVID=99282250