Fix docker_build file path when file is generated
If the file is generated, by a genrule for instance, then
the stripping of the prefix was incorrect because the used path
was containing 'bazel-out/...' prefix. This change moved the prefix
stripping to the Skylark rule so we can use "short_path".

Also fix the test paths.
Tested by bazel test //tools/build_defs/docker:build_test.

--
Change-Id: Ia3eb98268505002bade1f6b0dd916e1a9767a530
Reviewed-on: https://bazel-review.googlesource.com/1770
MOS_MIGRATED_REVID=99710676