Remove the some calls to AnalysisEnvironment.getDerivedArtifact() and add a big fat warning that that method shouldn't be used anymore.
Ideally, I'd mark it as deprecated, but it's legitimately used by RuleContext. I could just forbid access to AnalysisEnvironment to rule implementations, but, sadly, almost every method on it is called at a few oddball places in implementations of various rules, so we can't do it just yet.

--
MOS_MIGRATED_REVID=101345484