Always create a generating action for _proguard.jar, even if no specs are specified.
Because configurable attributes can be used to create an android_binary which has no
proguard_specs, and because it's impossible to tell this case at the time the implicit
outputs are defined, some generating action must be defined even if no proguard_specs
actually make it to the rule. If the proguard_generate_mapping attribute is specified,
the mapping is also generated. Both are generated by a FailAction and so are not
actually produced.

--
MOS_MIGRATED_REVID=101694294