Make android_binary work with sandboxing. Note that "bazel mobile-install" is still broken owing to a bug in the code of the sandbox (#413)
The alternative to hand-listing the dependencies of apkbuilder is to glob all the jars in tools/lib. It would be more robust, but would take a little more time in the loading phase and on the startup of apkbuilder. Let's see how far we get with this approach.

Quite an embarrassing omission, owing to the lack of integration tests, which was in turn caused by us not having a continuous build. Now that we have one, we can actually write an integration test.

--
MOS_MIGRATED_REVID=102123234