Fix a signed vs unsigned comparison warning in the ijar zipper.
This is more than just a cosmetic issue. Since groovy and appengine
rules depend directly on this target, the zipper gets built in the
user's workspace whenever a groovy/appengine target is built. If the
workspace is set to upgrade warnings to errors, this can cause builds
to fail.

--
Change-Id: I6c8b347df14098945c788411fc3e38f9c128596f
Reviewed-on: https://bazel-review.googlesource.com/1951
MOS_MIGRATED_REVID=102223767