Quick fix for shipping the android tools in the Bazel binary
This is not really nice, there are several hacks there. This
repository will get removed in the future and linked to a
remote one instead. I tested it against the tutorial and
it works like a charm. mobile-install seems like to work,
maybe that's also fixing the last issue reported in #392.

Known issue: Java compilation output errors about files being
modified in the future.

--
MOS_MIGRATED_REVID=102282979