Partially revert change to collect baseline coverage from individual libraries.
We still want to do that, but not like this. Our infrastructure supports
per-target coverage, and so we also need to support per-target baseline
coverage.

I'm working on better documentation (not hard to be better than no docs),
which will cover this. I left a couple of TODOs to explain how we want to do
it in the future.

--
MOS_MIGRATED_REVID=103379710