Remove BuildView.skyframeCacheWasInvalidated.
It was mainly used for testing during the transition phase to Skyframe. The
only reason I can see to keep it would be to have test coverage that changing
the configuration drops the configured targets. However, do we even want that?
I tried to find out whether it's safe to remove that check (the corresponding
comment is outdated), but couldn't find where (if anywhere) we're doing
garbage collection of old configured targets.

--
MOS_MIGRATED_REVID=104009210