Incremental build and load of Docker layers
Previously we were always tar all the layers in the final image.
This change do not build the full image except if required. It instead
only build the concerned layer and a loader that loads all the layer
that were not previously loaded in your local registry.

This is somewhat incompatible with the previous version since the full
tar file is no more build if not explicitely requested.

RELNOTES: [docker_build] incremental loading is default now.
          Specify explicitly //package:target.tar (with the .tar extension)
          to obtain the full image.

--
Change-Id: I75c96d819f2a0f7c9444723b04b660b4939350ec
Reviewed-on: https://bazel-review.googlesource.com/#/c/1820/
MOS_MIGRATED_REVID=104737560