Make a copy of the list when passing the value from Skylark to BUILD.
This change affects only lists, because it is the only mutable type.
The change is also temporary. We will later make the list immutable (when
it comes from a different context) and we will be able to remove the copy.

This fixes a critical bug where values could be mutated from a BUILD file
and affect other BUILD files.

--
MOS_MIGRATED_REVID=105400506