Rollback of commit 97eaf9136abad30827cf1a060cc32e786745923d.
*** Reason for rollback ***

Broke bazel on mac

*** Original change description ***

Add a //src:bazel-notools target that is the same as the Bazel binary except without the embedded tools.

Its use requires the addition of a local_workspace(name = "bazel_tools", path=...) to the WORKSPACE file. It is useful if you don't care about the set of embedded tools (because you provide them to Bazel in another way anyway) but you do care about the size of the Bazel binary (it's 16 MB compared to 56 MB with the tools).

--
MOS_MIGRATED_REVID=105553886