Fix for SIOCGIFINDEX undefined on Mac (https://github.com/bazelbuild/bazel/issues/510)
This compiles with the assumption that all we were using SIOCGIFINDEX for was to verify
that the name was valid. I think the code as written before was actually wrong as the
index value, and the flags value for a ifreq structure actually shared memory, so we were
potentially setting flags that we didn't want to set based on the actual index returned.

--
MOS_MIGRATED_REVID=105619856