In SkyQueryEnvironment, check TransitiveTraversalValues for errors
In SkyQueryEnvironment, the buildTransitiveClosure method ignored the
error messages in successfully evaluated TransitiveTraversalValues,
under the belief that any such error messages are also obtainable by
calling getMissingAndExceptions. This happens to be true right now,
but may not be true in the future.

If, in particular, the TransitiveTraversal skyfunction inlines its
call to the TargetMarker skyfunction, those error messages won't be
available through getMissingAndExceptions.

--
MOS_MIGRATED_REVID=105980430