Fix broken / backwards logic in the blaze launcher when trying to kill a running server (because startup flags changed, different Blaze version, etc.). Currently, we do a SIGTERM then fallback to a SIGKILL. Our logic to check if the server was still up after the fallback was backwards, so in some cases we would connect to the old server even though we just sent it a SIGKILL. Not really ideal.
Note that this only occurred if the SIGTERM failed to kill the Blaze process within 10s, so this failure mode happened rarely.

Also: Check for the expected errno from the kill(), and exit with an "internal error" exit code instead of "1" if we are unable to kill the old server.

--
MOS_MIGRATED_REVID=105999230